Renamed classes folder to 'groups' to lessen ambiguity
